Fix: IDENTITY columns must be BIGINT in Fabric Warehouse
Input.Targets.pk_target and Gold.Fact_Targets.bk_target_id: INT -> BIGINT

+++ Fabric/Gold.Fact_Targets.Table.sql
@@ -8,7 +8,7 @@ GO
 CREATE TABLE [Gold].[Fact_Targets] (
     [pk_target]          BIGINT IDENTITY NOT NULL,
     [Tenant_ID]          INT          NOT NULL,
-    [bk_target_id]       INT          NOT NULL,
+    [bk_target_id]       BIGINT       NOT NULL,
     [fk_Practice_Site]   BIGINT       NOT NULL,
     [fk_Practitioner]    BIGINT       NOT NULL,
     [fk_Date]            BIGINT       NOT NULL,
